implementasi sistem login dan register serta validasi data berdasarkan user

+++ app_dashboard/app/Http/Controllers/StudyDayController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\StudyDay;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class StudyDayController extends Controller
 {
@@ -12,7 +14,7 @@ class StudyDayController extends Controller
      */
     public function index()
     {
-        $all_schedules = StudyDay::all() ?? [];
+        $all_schedules = Auth::user()->studyDays ?? [];
 
         view('layout.layout', compact('all_schedules'));
         
@@ -37,11 +39,11 @@ class StudyDayController extends Controller
                 'date' => ['required', 'date'],
             ]);
 
-            StudyDay::create([
-                'date' => $request->date,
+            Auth::user()->studyDays()->create([
+                'date' => $request->date
             ]);
 
-            return redirect("/")
+            return redirect("/study-day")
                 ->with('success', 'Schedule created successfully.');
        }
        catch (\Throwable $th) {
@@ -52,18 +54,12 @@ class StudyDayController extends Controller
     }
 
     /**
-     * Display the specified resource.
-     */
-    public function show(StudyDay $studyDay)
-    {
-        //
-    }
-
-    /**
      * Show the form for editing the specified resource.
      */
     public function edit(StudyDay $studyDay)
     {
+        Gate::authorize('update', $studyDay);
+
         return view('dashboard.study-day.form_edit', compact('studyDay'));
     }
 
@@ -72,16 +68,18 @@ class StudyDayController extends Controller
      */
     public function update(Request $request, StudyDay $studyDay)
     {
+        Gate::authorize('update', $studyDay);
+        
         try {
             $request->validate([
                 'date' => ['required', 'date'],
             ]);
 
             $studyDay->update([
-                'date' => $request->date,
+                'date' => $request->date
             ]);
 
-            return redirect("/")
+            return redirect("/study-day")
                 ->with('success', 'Schedule updated successfully.');
         }
         catch (\Throwable $th) {
@@ -96,10 +94,12 @@ class StudyDayController extends Controller
      */
     public function destroy(StudyDay $studyDay)
     {
+        Gate::authorize('delete', $studyDay);
+
         try{
             $studyDay->delete();
         
-            return redirect("/")
+            return redirect("/study-day")
                 ->with('success', 'Schedule deleted successfully.');
         }
         catch (\Throwable $th) {
